% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_theme.R
\name{theme_bw_new}
\alias{theme_bw_new}
\title{theme_bw_new}
\usage{
theme_bw_new(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_size}{numeric font size}

\item{base_family}{character font family}

\item{base_line_size}{numeric}

\item{base_rect_size}{numeric}
}
\description{
new bw theme based on ggplot2
}
\examples{
library(ggplot2)
raw_plot <- ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, color = Species)) + geom_point()
raw_plot
raw_plot + theme_bw_new()

}
